fix: Add API Operation
