+++ SeamlessPayCore/Classes/APIClient/APIClient.swift
@@ -340,48 +340,7 @@ private extension APIClient {
 
 // MARK: API operation
 private extension APIClient {
-  enum APIOperation {
-    case createToken
-    case createCustomer
-    case updateCustomer(id: String)
-    case retrieveCustomer(id: String)
-    case createCharge
-    case retrieveCharge(id: String)
-    case listCharges
-
-    var path: String {
-      let body: String
-      switch self {
-      case .createToken:
-        body = "tokens"
-      case .createCustomer:
-        body = "customers"
-      case let .retrieveCharge(value),
-           let .retrieveCustomer(value),
-           let .updateCustomer(value):
-        body = ["customers", value].joined(separator: "/")
-      case .createCharge,
-           .listCharges:
-        body = "charges"
-      }
-      return "/" + body
-    }
-
-    var method: HTTPMethod {
-      switch self {
-      case .createCharge,
-           .createCustomer,
-           .createToken:
-        return .post
-      case .updateCustomer:
-        return .put
-      case .listCharges,
-           .retrieveCharge,
-           .retrieveCustomer:
-        return .get
-      }
-    }
-  }
+  
 }
 
 // MARK: Helpers
